+++ xcube_resampling/gridmapping/coords.py
@@ -296,8 +296,12 @@ def new_grid_mapping_from_coords(
     x_res, y_res = _to_int_or_float(x_res), _to_int_or_float(y_res)
     if xy_bbox is None:
         x_res_05, y_res_05 = x_res / 2, y_res / 2
-        x_min = _to_int_or_float(x_coords[..., 0].min() - x_res_05)
-        x_max = _to_int_or_float(x_coords[..., -1].max() + x_res_05)
+        if np.any(x_coords[..., 0] > x_coords[..., -1]):
+            x_min = _to_int_or_float(x_coords[..., -1].min() + x_res_05)
+            x_max = _to_int_or_float(x_coords[..., 0].max() - x_res_05)
+        else:
+            x_min = _to_int_or_float(x_coords[..., 0].min() - x_res_05)
+            x_max = _to_int_or_float(x_coords[..., -1].max() + x_res_05)
         if is_j_axis_up:
             y_min = _to_int_or_float(float(y_coords[0, ...].min()) - y_res_05)
             y_max = _to_int_or_float(float(y_coords[-1, ...].max()) + y_res_05)